requirements.txt
# Automatically generated by gen_requirements_all.py, do not edit

# -c homeassistant/package_constraints.txt

# Home Assistant Core
aiodns==3.5.0
aiohasupervisor==0.3.2
aiohttp==3.12.15
aiohttp_cors==0.8.1
aiohttp-fast-zlib==0.3.0
aiohttp-asyncmdnsresolver==0.1.1
aiozoneinfo==0.2.3
annotatedyaml==0.4.5
astral==2.2
async-interrupt==1.2.2
attrs==25.3.0
atomicwrites-homeassistant==1.4.1
audioop-lts==0.2.1
awesomeversion==25.5.0
bcrypt==4.3.0
certifi>=2021.5.30
ciso8601==2.3.3
cronsim==2.6
fnv-hash-fast==1.5.0
hass-nabucasa==1.1.1
httpx==0.28.1
home-assistant-bluetooth==1.13.1
ifaddr==0.2.0
Jinja2==3.1.6
lru-dict==1.3.0
PyJWT==2.10.1
cryptography==45.0.7
Pillow==11.3.0
propcache==0.3.2
pyOpenSSL==25.1.0
orjson==3.11.3
packaging>=23.1
psutil-home-assistant==0.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.5
securetar==2025.2.1
SQLAlchemy==2.0.41
standard-aifc==3.13.0
standard-telnetlib==3.13.0
typing-extensions>=4.15.0,<5.0
ulid-transform==1.4.0
urllib3>=2.0
uv==0.8.9
voluptuous==0.15.2
voluptuous-serialize==2.7.0
voluptuous-openapi==0.1.0
yarl==1.20.1
webrtc-models==0.3.0
zeroconf==0.147.2
